requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Django==4.2.6
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
logger==1.4
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.0
Pillow==9.5.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.200
urllib3==2.0.7
win32-setctime==1.1.0
